+++ R/design.R
@@ -1,4 +1,6 @@
 obtain.design <- function(data, formula = ~.) {
+  # try out the following
+  # formula <- update(formula, . ~ . - 1)
   mf <- model.frame(formula, data = data, na.action = na.pass)
 
   # Convert logical variables to numeric to prevent dummy expansion